python-backend/requirements.txt
# Core
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
pyspellchecker
httpx
python-dotenv
psutil
python-multipart

# Speech Recognition (existente)
SpeechRecognition
sounddevice
soundfile

# LLM - OpenRouter
openai

# STT - emotion2vec (pesado, requiere 8+ GB RAM)
transformers
torch
librosa

# STT - Vosk (ligero, fallback)
vosk

# TTS - Kokoro (local, neural, alta calidad)
kokoro

# Utilidades
numpy
